+++ skins/50-shades-of-noah-grey/skin-footer-scripts.php
@@ -1,6 +1,7 @@
 <?php
 /**
  * SNAPSMACK - Footer scripts for the 50-shades-of-noah-grey skin
+ * Alpha v0.7.9c
  *
  * Reserved for future use. Core scripts load globally.
  */